backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
groq>=0.12.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
scikit-learn>=1.6.0
python-multipart>=0.0.17
